feat:opt cmd
