+++ docs/zh/快速启用.md
@@ -307,12 +307,12 @@ addr: /ip4/192.168.19.67/tcp/58418
 ### 存储挂单
 
 ```bash
-./venus-market storage-deals set-ask --miner=t01041 --price=0.01fil --verified-price=0.02fil --min-piece-size=512b --max-piece-size=512M
+./venus-market storage ask set --price=0.01fil --verified-price=0.02fil --min-piece-size=512b --max-piece-size=512M t01041
 ```
 可以通过命令行工具检查定价信息：
 
 ```shell
-./venus-market storage-deals get-ask --miner=t01041
+./venus-market storage ask get t01041
 Price per GiB/Epoch  Verified  Min. Piece Size (padded)  Max. Piece Size (padded)  Expiry (Epoch)  Expiry (Appx. Rem. Time)  Seq. No.
 0.01 FIL             0.02 FIL  512 B                     521 MiB                     161256          719h59m0s                 0
 ```
@@ -322,16 +322,16 @@ Price per GiB/Epoch  Verified  Min. Piece Size (padded)  Max. Piece Size (padded
 存储服务提供商至少应设置收款地址
 
 ```bash
-./venus-market retrieval-deals set-ask --payment-addr t3ueb62v5kbyuvwo5tuyzpvds2bfakdjeg2s33p47buvbfiyd7w5fwmeilobt5cqzi673s5z6i267igkgxum6a
+./venus-market retrieve ask set t3ueb62v5kbyuvwo5tuyzpvds2bfakdjeg2s33p47buvbfiyd7w5fwmeilobt5cqzi673s5z6i267igkgxum6a
 ```
 
 同时，也可以设置数据检索订单的价格，不设置时，默认为0.
 ```bash
-./venus-market retrieval-deals set-ask \
+./venus-market retrieve ask set \
 --price 0.02fil \
 --unseal-price 0.01fil \
 --payment-interval 1MB \
---payment-addr t3ueb62v5kbyuvwo5tuyzpvds2bfakdjeg2s33p47buvbfiyd7w5fwmeilobt5cqzi673s5z6i267igkgxum6a
+t3ueb62v5kbyuvwo5tuyzpvds2bfakdjeg2s33p47buvbfiyd7w5fwmeilobt5cqzi673s5z6i267igkgxum6a
 ```
